+++ qCC/mainwindow.h
@@ -431,6 +431,10 @@ private slots:
 	//! Creates a cloud with the (bounding-box) centers of all selected entities
 	void doActionCreateCloudFromEntCenters();
 
+	inline void doActionMoveBBCenterToOrigin()    { doActionFastRegistration(MoveBBCenterToOrigin); }
+	inline void doActionMoveBBMinCornerToOrigin() { doActionFastRegistration(MoveBBMinCornerToOrigin); }
+	inline void doActionMoveBBMaxCornerToOrigin() { doActionFastRegistration(MoveBBMaxCornerToOrigin); }
+
 private:
 	//! Shortcut: asks the user to select one cloud
 	/** \param defaultCloudEntity a cloud to select by default (optional)
@@ -439,8 +443,17 @@ private:
 	**/
 	ccPointCloud* askUserToSelectACloud(ccHObject* defaultCloudEntity = nullptr, QString inviteMessage = QString());
 
-	void	toggleSelectedEntitiesProperty( ccEntityAction::TOGGLE_PROPERTY property );
-	void	clearSelectedEntitiesProperty( ccEntityAction::CLEAR_PROPERTY property );
+	enum FastRegistrationMode
+	{
+		MoveBBCenterToOrigin,
+		MoveBBMinCornerToOrigin,
+		MoveBBMaxCornerToOrigin
+	};
+
+	void doActionFastRegistration(FastRegistrationMode mode);
+
+	void toggleSelectedEntitiesProperty( ccEntityAction::TOGGLE_PROPERTY property );
+	void clearSelectedEntitiesProperty( ccEntityAction::CLEAR_PROPERTY property );
 	
 	void setView( CC_VIEW_ORIENTATION view ) override;